Translated to FastAPI and minor fixes/chanes

+++ api/unittests.py
@@ -1,43 +1,71 @@
 #!/usr/bin/python3
 import unittest
-from requests import get, post, put
-from threading import Thread
-from time import sleep
-from devapi import main as server_main
-from os import mkdir
-from shutil import rmtree  # NOTE: To reduce bloat, we need to get rid of this
-
-# ALERT: Set this to True for debugging
-DEBUG = False
+from dataclasses import dataclass, field
+# ALERT: Set to true for performance test
+PERF = False
+# All it does is not load our Kerplunk backend,
+# and thus assumes that our backend is already running
 
 # ALERT: When writing new unittests, pay attention to the name's number
 # TestUser's data
 supported_fields = ('uname', 'lname')
 
-
-class User():
-    def __init__(self, un: str, uhash: str, email: str,
-                 salt: str, sid: str, ln: str):
-        self.uname: str = un
-        self.hash: str = uhash
-        self.email: str = email
-        self.salt: str = salt
-        self.sid: str = sid
-        self.lname: str = ln
+if PERF:
+    from httpx import get, post, put, delete
+else:
+    from fastapi.testclient import TestClient
+    from main import app
+    from os import mkdir
+    from shutil import rmtree
+    # NOTE: To reduce bloat, we need to get rid of this
 
 
-class Chat():
-    def __init__(self, n: str, p: str = '', o: str = '',
-                 port: int = 0, ps: list[str] = []):
-        self.name: str = n
-        self.pwd: str = p
-        self.port: int = port
-        self.owner: str = o
-        self.participants: list[str] = ps
+class performance_test_client():
+    def __init__(self, url: str):
+        self.url = url
+
+    def get(self, url: str, params={}, headers={}):
+        return get(self.url + url, params=params, headers=headers)
+
+    def post(self, url: str, data={}, json={}, headers={}):
+        return post(self.url + url, data=data, json=json, headers=headers)
+
+    def patch(self, url: str, json: dict, params: dict, headers: dict):
+        return put(self.url + url, json=json, params=params, headers=headers)
+
+    def delete(self, url: str, params: dict):
+        return delete(self.url + url, params=params)
 
 
-user1: User = User('test', 'abcd', 'test@test.com', 'efgh',
-                   '', 'Newton,Newtonson')
+def make_client():
+    if not PERF:
+        return TestClient(app)
+    else:
+        return performance_test_client('http://127.0.0.1:8000')
+
+
+client = make_client()
+
+
+@dataclass
+class User:
+    uname: str
+    pwd: str
+    email: str
+    sid: str
+    lname: str
+
+
+@dataclass
+class Chat:
+    name: str
+    pwd: str = ''
+    cid: str = ''
+    owner: str = ''
+    participants: list = field(default_factory=lambda: [])
+
+
+user1: User = User('test', 'abcd', 'test@test.com', '', 'Newton,Newtonson')
 # alt vals is used in Test02-test110
 # Whenever supported_fields is updated,
 # this should also be updated (On New Features)
@@ -47,146 +75,103 @@ altvals = {
 }
 
 # This is a second user
-user2: User = User('test2', 'seconduser',
-                   'test2@test.com', 'zxcvbn', '', 'Sleepy,Sleeper')
+user2: User = User('test2', 'seconduser', 'test2@test.com', '', 'Slepy,Sleper')
 users = (user1, user2)
 chat1: Chat = Chat('Testing')
 chat_rooms = (chat1, )
 
 
-def url(addr):
-    return 'http://127.0.0.1:8000/api/' + addr
-
-
-def sget(addr: str, params: dict = {}):
-    r = get(url=url(addr), params=params)
-    if DEBUG:
-        print('GET to ' + addr + ' returned:\n' + str(r.content))
-    return r.json()
-
-
-def spost(addr: str, form_data: dict):
-    r = post(url=url(addr), json=form_data)
-    if DEBUG:
-        print('POST to ' + addr + ' returned:\n' + str(r.content))
-    return r.json()
-
-
-def sput(addr: str, data: dict):
-    r = put(url=url(addr), data=data)
-    if DEBUG:
-        print('PUT to ' + addr + ' returned:\n' + str(r.content))
-    return r.json()
+def genhed(user: User):
+    return {
+        'Authorization': 'Bearer ' + user.sid
+    }
 
 
 def sign_up_user(user: User):
-    fd = {
-        'username': user.uname,
-        'hash': user.hash,
-        'email': user.email,
-        'salt': user.salt
-    }
-    return spost('auth', fd)['error']
-
-
-def get_user_salt(user: User):
-    p = {
-        'username': user.uname
-    }
-    return sget('auth?', p)['salt']
+    return client.post('/signup', json={
+        'uname': user.uname,
+        'pwd': user.pwd,
+        'email': user.email
+    })
 
 
 def get_user_sid(user: User):
-    p = {
+    con = {
+        'grant_type': '',
         'username': user.uname,
-        'hash': user.hash
+        'password': user.pwd,
+        'scope': '',
+        'client_id': '',
+        'client_secret': ''
     }
-    return sget('auth?', p)['sid']
+    return client.post('/login', data=con).json()['access_token']
 
 
 class Test010_auth(unittest.TestCase):
     def test010_make_user(self):
         """Sign up test user"""
-        error = sign_up_user(user1)
-        self.assertFalse(error)
+        self.assertEqual(sign_up_user(user1).status_code, 200)
 
-    def test020_get_salt(self):
-        """Get Back Salt"""
-        salt = get_user_salt(user1)
-        self.assertEqual(salt, user1.salt)
-
-    def test030_get_sid(self):
+    def test020_get_sid(self):
         """
         Generate an SID, Must be done after we
         get the salt in test020_get_salt above
         """
         sid = get_user_sid(user1)
-        self.assertTrue(sid)
+        self.assertIsNotNone(sid)
         user1.sid = sid
 
-    def test040_login_second_user(self):
+    def test030_login_second_user(self):
         sign_up_user(user2)
-        get_user_salt(user2)  # We don't need the salt in this case
         sid = get_user_sid(user2)
-        self.assertTrue(sid)
+        self.assertIsNotNone(sid)
         user2.sid = sid
 
 
 class Test020_User_Data(unittest.TestCase):
     def test010_get_username(self):
         """Check if test user's username is correctly registered"""
-        p = {
-            'sid': user1.sid,
-            'fields': 'uname'
-        }
-        r = sget('users?', p)
-        self.assertTrue(r['uname'])
+        r = client.get('/userme', params={
+            'fields': ['uname']
+            }, headers=genhed(user1)).json()
+        self.assertIn('uname', r)
         self.assertEqual(r['uname'], user1.uname)
 
     def test020_set_legalname(self):
         """Set lname value in db"""
-        fd = {
-            'sid': user1.sid,
-            'lname': user1.lname
-        }
-        r = spost('users', fd)
+        r = client.post('/userme', json={'lname': user1.lname},
+                        headers=genhed(user1)).json()
+        self.assertIn('changed', r)
         self.assertEqual(r['changed'], 'lname')
 
     def test030_get_legalname(self):
         """Check if lname value in db is correct"""
-        p = {
-            'sid': user1.sid,
+        r = client.get('/userme', params={
             'fields': 'lname'
-        }
-        r = sget('users?', p)
+            }, headers=genhed(user1)).json()
         self.assertEqual(r['lname'], user1.lname)
 
-    def test100_multiget(self, key=user1):
+    def test100_multiget(self, key=user1, auth_user=user1):
         """
         Check if getting all values at once works.
         The key is what will be used as benchmark for the returned values.
         test120_check_multipost will set key to altvals
         """
-        p = {
-            'sid': user1.sid,
-            'fields': str(supported_fields)
-                        .replace(' ', '').replace('(', '')
-                        .replace(')', '').replace("'", '')
-        }
-        r = sget('users?', p)
-        if isinstance(key, dict):
-            key = key.__dict__  # FORM INTO DICT
+        # We need key to be a dictionary, eventually
+        if isinstance(key, User):
+            key = key.__dict__
+        r = client.get('/userme', params={
+            'fields': supported_fields
+            }, headers=genhed(auth_user)).json()
         for f in supported_fields:
             # Class.__dict__ is a quick way to get it done, apparently
             self.assertEqual(r[f], key[f])
 
     def test110_multipost(self):
         """Change all values"""
-        fd = {'sid': user1.sid}
-        for k, v in altvals.items():
-            fd[k] = v
-        r = spost('users', fd)
-        self.assertEqual(set(supported_fields), set(r['changed'].split(' ')))
+        r = client.post('/userme', json=altvals, headers=genhed(user1)).json()
+        self.assertIn('changed', r)
+        self.assertSetEqual(set(supported_fields), set(r['changed'].split(' ')))
 
     def test120_check_multipost(self):
         """Check if all new values are properly set, after 110_multipost"""
@@ -198,52 +183,49 @@ class Test020_User_Data(unittest.TestCase):
 class Test030_Chat_Rooms(unittest.TestCase):
     def test010_open_chat_room(self):
         """Check if API can create chat rooms"""
-        fd = {
-            'sid': user1.sid,
-            'name': chat_rooms[0].name,
-            'pwd': chat_rooms[0].pwd,
+        chat: Chat = chat_rooms[0]
+        js = {
+            'name': chat.name,
+            'pwd': chat.pwd,
             'public': True,
             'temp': True
         }
-        r = spost('chatrooms', fd)
-        self.assertFalse(r['error'])
+        r = client.post('/chats', json=js, headers=genhed(user1)).json()
+        self.assertIn('cid', r)
+        chat.cid = r['cid']
 
     def test020_list_chat_rooms(self):
         """Check if API returns correct list of chat rooms"""
         # No SID required because this doesn't really need authentication
         # And no parameters required
-        r = sget('chatrooms')
-        self.assertEqual(r['chats'], chat_rooms[0].name)
+        r = client.get('/chats').json()
+        self.assertEqual(r['chatrooms'], chat_rooms[0].name)
 
     def test030_first_user_join_chat(self):
         """Check what data is being given back from the
         first time a user joins the chat"""
         chat: Chat = chat_rooms[0]
-        d = {
-            'sid': user2.sid,
+        r = client.patch('/chats', params={
             'name': chat_rooms[0].name
-        }
-        r = sput('chatrooms', d)
-        self.assertFalse(r['error'])
-        self.assertEqual(r['port'], 6000)
-        # Should be the very first port, & we start from 6000
-        self.assertMultiLineEqual(r['owner'], user1.uname)
+            }, headers=genhed(user2)).json()
+        self.assertEqual(r['owner'], user1.uname)
+        self.assertEqual(r['cid'], chat.cid)
         self.assertListEqual(r['joiners'], [user2.uname])
         chat.owner = r['owner']
-        chat.port = r['port']
         chat.pwd = ''
 
 
 if __name__ == '__main__':
-    # Clear data directory
-    rmtree('../data/data/')
-    # NOTE: We need to eventually implement our own rmtree,
-    # because it's a waste of bandwidth installing a whole library
-    # to do something relatively simple, recursively removing a directory
-    mkdir('../data/data/')
+    if not PERF:
+        # Clear data directory
+        rmtree('../data/data/')
+        # NOTE: We need to eventually implement our own rmtree,
+        # because it's a waste of bandwidth installing a whole library
+        # to do something relatively simple, recursively removing a directory
+        mkdir('../data/data/')
     # Start Server as daemon (dies when script ends) on seperate thread
-    server = Thread(target=server_main)
-    server.daemon = True
-    server.start()
-    sleep(1)  # To give server time to start
+    # server = Thread(target=server_main)
+    # server.daemon = True
+    # server.start()
+    # sleep(1)  # To give server time to start
     unittest.main()
